Fixed some base errors, enabled for cygwin compile

+++ src/Data/DoubleList.cpp
@@ -122,7 +122,7 @@ double DoubleList::pop(){
 };
 
 int DoubleList::size(){
-    return _Size;
+    return (int)_Size;
 }
 
 void DoubleList::clear(){
